Simplified example app by removing project files Moved generate_secure_signature to utils package Addressed some review comments
