+++ examples/signeduploads/sampleapp/forms.py
@@ -1,5 +1,6 @@
 from django import forms
 from pyuploadcare.dj.forms import ImageGroupField
 
+
 class ImgForm(forms.Form):
     picImg = ImageGroupField(label='')
