--- examples/signeduploads/sampleapp/models.py
@@ -1,5 +1,0 @@
-from __future__ import unicode_literals
-
-from django.db import models
-
-# Create your models here.
