+++ examples/signeduploads/sampleapp/urls.py
@@ -4,6 +4,6 @@ from . import views
 
 urlpatterns = [
     url(r'index$', views.index, name='index'),
-    url(r'release_signature$', views.release_signature, name='release_signature'),
+    url(r'release_signature$',
+        views.release_signature, name='release_signature'),
 ]
-
