+++ pyuploadcare/conf.py
@@ -14,6 +14,7 @@ DEFAULT = object()
 
 pub_key = None
 secret = None
+expire = None
 signed_uploads = None
 
 api_version = '0.5'
